Added a software display class
It uses the curses terminal UI library. It presents a two-digit display in the
terminal with the same API as the hardware TwoDigitDisplay.

This will let me hack on the scoreboard logic without having to deploy the
code to the RPi with the LED display attached. I can run it all directly on
the laptop I'm coding from.

In a way it's an emulator for my hardware board.

Also, I moved the sample logic out into a separate displaysample.py. It can
use either the hardware or software display to run the silly counting demo.