Fix a fundamental issue in bit shifting
I was apparently doing latching and clocking in the wrong order. This manifest
itself when I was trying to use the SingleDigitDisplay to set a single solid
digit (without pulsing). It was briefly lighting up the right segments and
then lighting them all.

Doing the latching and clocking in this new order seems to have fixed the
issue.