Adds the start of better in-game state tracking
Now that the scoreboard shows balls, strikes and outs it can wind up handling
some states that it doesn't represent well. Like, 2 balls, 3 strikes, 1 out
(where the 1 out is by virtue of the 3 strikes). Really the scoreboard should
display the out and no balls and strikes in preparation for the next batter
coming to the plate. Same thing with innings - once 3 outs are recorded, the
inning is over and the next inning (or half inning) should be shown.